Refactor getStats methods in RTCPeerConnection, SecureTransportManager, and TransceiverManager for improved stats collection and organization
